Add i2c to gpio pins of your choice.
Set them up as usual with gpio-mode, use
    gpio-mode i2c sda
    gpio-mode i2c scl
to select the data and clock lines and reset.

Then you can use the commands from port tcp port 24:
    i2c-address <address>	set slave address (in hex, no 0x)
    i2c-send <byte> <byte> ...	send bytes to slave (in hex, no 0x)
    i2c-receive <amount>	retrieve this number of bytes from the slave
    i2c-reset		use this to reset the bus state after an error
			occured, it will try to get all slaves into a
			sane condition as well.

This is quite preliminary, I only tested it on a tmp275 at 3.3V.

+++ stats.c
@@ -32,7 +32,7 @@ static const char *reset_map[] =
 	"unknown"
 };
 
-ICACHE_FLASH_ATTR void stats_generate(size_t size, char *dst)
+ICACHE_FLASH_ATTR void stats_generate(uint16_t size, char *dst)
 {
 	const struct rst_info *rst_info;
